* Add check if dir or file exist before doing operation. * Add Remove /tmp/spksrc dir if exist on start of script.

+++ buildAndPatch.sh
@@ -1,12 +1,16 @@
+# remove old nginx build dir
+[[ -d /tmp/spksrc ]] && rm -r /tmp/spksrc
 # Temp dir for the nginx build
 mkdir /tmp/spksrc
 # Create dir for user settings: reverse proxy,
-# theese copied to app.d when creating a reverse proxy.
-mkdir /etc/nginx/reverseproxy/
+# theese copied to app.d when creating a reverse proxy
+[[ ! -d /etc/nginx/reverseproxy/ ]] && mkdir /etc/nginx/reverseproxy/
+
 # link our libpcreso ssl works.
-ln -s /usr/lib/libpcre.so /usr/lib/libpcre.so.3
+[[ ! -e /usr/lib/libpcre.so.3 ]] && ln -s /usr/lib/libpcre.so /usr/lib/libpcre.so.3
+
 # take owenrship of dir as http.
-chown http:http /etc/nginx/scgi_temp
+[[ -d /etc/nginx/scgi_temp/ ]] && chown http:http /etc/nginx/scgi_temp
 
 # Build nginx
 docker run -it --rm \
